refactor(public_www): rename reservation submit factory and split helpers
Replace useReservationSubmit (no hooks) with createReservationSubmitHandler.
Extract buildReservationSummary, buildReservationPayload, and trackSubmitError
into reservation-submit-helpers.ts with shared types in reservation-submit-types.ts.
